pyproject.toml
[project]
name = "docchatai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "cryptography>=46.0.3",
    "jinja2>=3.1.6",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph-cli[inmem]>=0.4.11",
    "pypdf>=6.5.0",
    "rapidocr-onnxruntime>=1.4.4",
    "sentence-transformers>=5.2.0",
    "sentencepiece>=0.2.1",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
